// src/config/enum/register.enum.ts
export enum REGISTER_ENUM {
  USERNAME = 'id',
  PASSWORD = 'password',
  CONFIRM_PASSWORD = 'confirm_password',
  DOB = 'dateOfBirth',
  EMAIL = 'email',
  PHONE_NUMBER = 'phone',
  CAPTCHA = 'captcha',
  FIRST_NAME = 'firstName',
  LAST_NAME = 'lastName',
  COUNTRY_ID = 'countryId',
  CURRENCY_ID = 'currencyId',
  JURISDICTION_ID = 'jurisdictionId',
  PLAYER_STATUS = 'playerStatus',
  CHARACTER_ID = 'characterId',
  PLAYER_LEVEL = 'playerLevel',
  REFERRAL_ID = 'referralId',
}